main: Another attempt at Secure Boot work-around
When Secure Boot is enabled, /sys/kernel/debug/vgaswitcheroo/switch
can't be opened, even read-only, but we can easily check for it being
available.

Check whether the file exists, and if it does, consider that we have
a dual GPU system. If the file can be opened, try to force the
integrated card to be used, but don't make failure to do that fatal.

Closes: #12